refactor: reorganize tensor utility functions and update imports for model optimization
