+++ test/unit/test_cpu/export/test_qlinear_pack_clamp.py
@@ -194,7 +194,7 @@ def test_missing_tensors_gptq_pack_clamps_zero_zero_point():
     """quantize_weight_rtn packs zp with the same (zp-1) convention; an
     all-positive weight yields zp=0 everywhere, which must clamp instead of
     smearing -1 across the packed zero-point words."""
-    from auto_round.utils.missing_tensors import quantize_weight_rtn
+    from auto_round.utils.model_free_utils import quantize_weight_rtn
 
     torch.manual_seed(0)
     bits, group = 4, 32
